using the variable `data`... it actually is already data, in the form of a JavaScript object, it is not related to a "notation" as in JavaScript Object Notation
using the variable `data`... it actually is already data, in the form of a JavaScript object, it is not related to a "notation" as in JavaScript Object Notation.  It actually gave some other developers or beginners to think a JavaScript object is "json" -- a notation.  JSON is more like, the JavaScript object, displayed as text or a string.